initial start on MarketModel2

+++ src/data/SystemData.java
@@ -621,7 +621,7 @@ public class SystemData {
 		if (fundamentalGenerator == null) {
 			return new Price(0);
 		}
-		return fundamentalGenerator.getValueAt((int) ts.longValue());
+		return fundamentalGenerator.getValueAt(ts);
 	}
 	
 	/**
